+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="be.ugent.zeus.hydra"
+<manifest xmlns:tools="http://schemas.android.com/tools"
+    package="be.ugent.zeus.hydra"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <!-- Network permissions -->
@@ -32,7 +33,9 @@
         android:label="@string/app_name"
         android:logo="@drawable/logo"
         android:roundIcon="@mipmap/ic_launcher_round"
-        android:theme="@style/Hydra.Main.NoActionBar">
+        android:theme="@style/Hydra.Main.NoActionBar"
+        tools:ignore="ManifestResource"
+        tools:targetApi="25">
 
         <!-- Main activity -->
         <activity
@@ -109,12 +112,6 @@
         <activity
             android:name=".activities.WebViewActivity"
             android:label="@string/title_activity_hydra_web_view"
-            android:parentActivityName=".activities.MainActivity" />
-
-        <!-- Show library view -->
-        <activity
-            android:name=".library.LibraryListActivity"
-            android:label="@string/title_activity_libraries"
             android:parentActivityName=".activities.MainActivity" />
 
         <!-- Show a Minerva course -->
