+++ app/src/main/res/menu/drawer_items.xml
@@ -10,17 +10,29 @@
             android:icon="@drawable/tabs_home"
             android:title="@string/title_home_feed" />
         <item
-            android:id="@+id/drawer_schamper"
-            android:icon="@drawable/tabs_schamper"
-            android:title="@string/title_home_schamper" />
-        <item
             android:id="@+id/drawer_resto"
             android:icon="@drawable/tabs_menu"
             android:title="@string/title_home_resto" />
         <item
+            android:id="@+id/drawer_minerva"
+            android:icon="@drawable/tabs_minerva"
+            android:title="@string/title_home_minerva" />
+        <item
             android:id="@+id/drawer_events"
             android:icon="@drawable/tabs_events"
             android:title="@string/title_home_events" />
+        <item
+            android:id="@+id/drawer_schamper"
+            android:icon="@drawable/tabs_schamper"
+            android:title="@string/title_home_schamper" />
+        <item
+            android:id="@+id/drawer_urgent"
+            android:icon="@drawable/tabs_urgent"
+            android:title="@string/title_home_urgent" />
+        <item
+            android:id="@+id/drawer_library"
+            android:icon="@drawable/tabs_library"
+            android:title="@string/title_library"/>
         <item
             android:id="@+id/drawer_news"
             android:icon="@drawable/tabs_news"
@@ -29,15 +41,6 @@
             android:id="@+id/drawer_info"
             android:icon="@drawable/tabs_info"
             android:title="@string/title_home_info" />
-        <item
-            android:id="@+id/drawer_minerva"
-            android:icon="@drawable/tabs_minerva"
-            android:title="@string/title_home_minerva" />
-
-        <item
-            android:id="@+id/drawer_urgent"
-            android:icon="@drawable/tabs_urgent"
-            android:title="@string/title_home_urgent" />
     </group>
 
     <group
